Reseting
I worked to finish the project so our feature/page will be taken. I know how it works, there will be compettition between teams to deliver. We won but it seemed like I won. The based for promotion is based on individual, not team.
This serves as a way to revert and help my team. WE ALL WILL BE IN 6.